create common superclass to the srg actions
