+++ src/main/kotlin/com/demonwav/mcdev/platform/mcp/actions/CopyAtAction.kt
@@ -13,29 +13,35 @@ package com.demonwav.mcdev.platform.mcp.actions
 import com.demonwav.mcdev.platform.mcp.srg.McpSrgMap
 import com.demonwav.mcdev.util.ActionData
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.psi.PsiClass
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiField
-import com.intellij.psi.PsiMethod
+import com.intellij.openapi.editor.Editor
+import com.intellij.psi.*
+import java.awt.Toolkit
+import java.awt.datatransfer.StringSelection
 
-class FindSrgMappingAction : SrgActionBase() {
-
+class CopyAtAction : SrgActionBase() {
     override fun withSrgTarget(parent: PsiElement, srgMap: McpSrgMap, e: AnActionEvent, data: ActionData) {
         when (parent) {
             is PsiField -> {
                 val srg = srgMap.findSrgField(parent) ?: return showBalloon("No SRG name found", e)
-                showSuccessBalloon(data.editor, data.element, "SRG name: "+srg.name)
+                copyToClipboard(data.editor, data.element, parent.containingClass?.qualifiedName+" "+srg.name+" #"+parent.name)
             }
             is PsiMethod -> {
                 val srg = srgMap.findSrgMethod(parent) ?: return showBalloon("No SRG name found", e)
-                showSuccessBalloon(data.editor, data.element, "SRG name: "+srg.name+srg.descriptor)
+                copyToClipboard(data.editor, data.element, parent.containingClass?.qualifiedName+" "+srg.name+srg.descriptor+" #"+parent.name)
             }
             is PsiClass -> {
                 val classMcpToSrg = srgMap.findSrgClass(parent) ?: return showBalloon("No SRG name found", e)
-                showSuccessBalloon(data.editor, data.element, "SRG name: "+classMcpToSrg)
+                copyToClipboard(data.editor, data.element, classMcpToSrg)
             }
             else -> showBalloon("Not a valid element", e)
         }
+    }
+
+    private fun copyToClipboard(editor: Editor, element: PsiElement, text: String) {
+        val stringSelection = StringSelection(text)
+        val clpbrd = Toolkit.getDefaultToolkit().systemClipboard
+        clpbrd.setContents(stringSelection, null)
+        showSuccessBalloon(editor, element, "Copied "+text)
     }
 
 }
