+++ src/main/kotlin/com/demonwav/mcdev/platform/mcp/actions/SrgActionBase.kt
@@ -11,7 +11,9 @@
 package com.demonwav.mcdev.platform.mcp.actions
 
 import com.demonwav.mcdev.platform.mcp.McpModuleType
+import com.demonwav.mcdev.platform.mcp.srg.McpSrgMap
 import com.demonwav.mcdev.platform.mixin.util.findFirstShadowTarget
+import com.demonwav.mcdev.util.ActionData
 import com.demonwav.mcdev.util.getDataFromActionEvent
 import com.demonwav.mcdev.util.invokeLater
 import com.intellij.openapi.actionSystem.AnAction
@@ -21,13 +23,15 @@ import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.ui.popup.Balloon
 import com.intellij.openapi.ui.popup.JBPopupFactory
 import com.intellij.openapi.wm.WindowManager
-import com.intellij.psi.*
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiIdentifier
+import com.intellij.psi.PsiMember
+import com.intellij.psi.PsiReference
 import com.intellij.ui.LightColors
 import com.intellij.ui.awt.RelativePoint
-import java.awt.Toolkit
-import java.awt.datatransfer.StringSelection
 
-class CopyATAction : AnAction() {
+
+abstract class SrgActionBase : AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
         val data = getDataFromActionEvent(e) ?: return showBalloon("Unknown failure", e)
 
@@ -52,25 +56,13 @@ class CopyATAction : AnAction() {
                 parent = parent.resolve()
             }
 
-            when (parent) {
-                is PsiField -> {
-                    val srg = srgMap.findSrgField(parent) ?: return@done showBalloon("No SRG name found", e)
-                    copyToClipboard(data.editor, data.element, parent.containingClass?.qualifiedName+" "+srg.name+" #"+parent.name)
-                }
-                is PsiMethod -> {
-                    val srg = srgMap.findSrgMethod(parent) ?: return@done showBalloon("No SRG name found", e)
-                    copyToClipboard(data.editor, data.element, parent.containingClass?.qualifiedName+" "+srg.name+srg.descriptor+" #"+parent.name)
-                }
-                is PsiClass -> {
-                    val classMcpToSrg = srgMap.findSrgClass(parent) ?: return@done showBalloon("No SRG name found", e)
-                    copyToClipboard(data.editor, data.element, classMcpToSrg)
-                }
-                else -> showBalloon("Not a valid element", e)
-            }
+            withSrgTarget(parent, srgMap, e, data)
         } ?: showBalloon("No mappings found", e)
     }
 
-    private fun showBalloon(message: String, e: AnActionEvent) {
+    abstract fun withSrgTarget(parent: PsiElement, srgMap: McpSrgMap, e: AnActionEvent, data: ActionData)
+
+    protected fun showBalloon(message: String, e: AnActionEvent) {
         val balloon = JBPopupFactory.getInstance()
             .createHtmlTextBalloonBuilder(message, null, LightColors.YELLOW, null)
             .setHideOnAction(true)
@@ -80,19 +72,10 @@ class CopyATAction : AnAction() {
 
         val statusBar = WindowManager.getInstance().getStatusBar(DataKeys.PROJECT.getData(e.dataContext))
 
-        invokeLater { 
-            balloon.show(RelativePoint.getCenterOf(statusBar.component), Balloon.Position.atRight) 
-        }
+        invokeLater { balloon.show(RelativePoint.getCenterOf(statusBar.component), Balloon.Position.atRight) }
     }
 
-    private fun copyToClipboard(editor: Editor, element: PsiElement, text: String) {
-        val stringSelection = StringSelection(text)
-        val clpbrd = Toolkit.getDefaultToolkit().systemClipboard
-        clpbrd.setContents(stringSelection, null)
-        showSuccessBalloon(editor, element, "Copied "+text)
-    }
-
-    private fun showSuccessBalloon(editor: Editor, element: PsiElement, text: String) {
+    protected fun showSuccessBalloon(editor: Editor, element: PsiElement, text: String) {
         val balloon = JBPopupFactory.getInstance()
             .createHtmlTextBalloonBuilder(text, null, LightColors.SLIGHTLY_GREEN, null)
             .setHideOnAction(true)
@@ -103,8 +86,8 @@ class CopyATAction : AnAction() {
         invokeLater {
             balloon.show(
                 RelativePoint(
-                        editor.contentComponent,
-                        editor.visualPositionToXY(editor.offsetToVisualPosition(element.textRange.endOffset))
+                    editor.contentComponent,
+                    editor.visualPositionToXY(editor.offsetToVisualPosition(element.textRange.endOffset))
                 ),
                 Balloon.Position.atRight
             )
