+++ src/main/resources/META-INF/plugin.xml
@@ -703,7 +703,7 @@
                 description="Find the associated SRG mapping for this element">
             <add-to-group relative-to-action="EditorPopupMenu2" anchor="after" group-id="EditorPopupMenu2"/>
         </action>
-        <action class="com.demonwav.mcdev.platform.mcp.actions.CopyATAction" id="CopyATAction"
+        <action class="com.demonwav.mcdev.platform.mcp.actions.CopyAtAction" id="CopyATAction"
                 text="Copy AT Entry to Clipboard"
                 description="Copy the reference to clipboard in Access Transformer format">
             <add-to-group relative-to-action="EditorPopupMenu2" anchor="after" group-id="EditorPopupMenu2"/>
